% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{normalize}
\alias{normalize}
\title{normalize}
\usage{
normalize(dfMatrix, vCost_Benefit, initCol = 1, initRow = 1)
}
\arguments{
\item{initCol}{First Column with values of input matrix (to avoid metadata), Default = 1}

\item{initRow}{First Row with values of input matrix (to avoid metadata), Default = 1}

\item{values}{A data set object with Alternatives X Criteria  values to be normalized}

\item{flags}{A vector of flags that determines if the criterion is a Cost or Benefit one
Must be same size of Criteria, must contains just strings initiated with B or C}
}
\value{
A data frame object that contains the input matrix with its values normalized
}
\description{
Normalize the values of Alternatives X Criteria matrix according to a given
             Cost-Benefit vector of flags
}
\examples{
normalize(dfMatrix, vCost_Benefit, initCol, initRow)
normalized_matrix <- normalize(row_values_matrix, flags_Cost_Benefit)
}
